+++ src-executables/Main-trident.hs
@@ -216,6 +216,7 @@ forgeOptParser = ForgeOptions <$> parseGenoDataSources
                               <*> parseMaybeOutPackageName
                               <*> parsePackageWise
                               <*> parseOutputPlinkPopMode
+                              <*> parseOutputOrdered
 
 genoconvertOptParser :: OP.Parser GenoconvertOptions
 genoconvertOptParser = GenoconvertOptions <$> parseGenoDataSources