+++ src/Poseidon/CLI/Forge.hs
@@ -78,6 +78,7 @@ data ForgeOptions = ForgeOptions
     , _forgeOutPacName         :: Maybe String
     , _forgePackageWise        :: Bool
     , _forgeOutputPlinkPopMode :: PlinkPopNameMode
+    , _forgeOutputOrdered      :: Bool
     }
 
 pacReadOpts :: PackageReadOptions
@@ -94,6 +95,7 @@ runForge (
                  entityInputs maybeSnpFile intersect_
                  outFormat minimal onlyGeno outPathRaw maybeOutName
                  packageWise outPlinkPopMode
+                 outputOrdered
     ) = do
 
     -- load packages --
@@ -138,7 +140,7 @@ runForge (
             else entities
 
     -- determine indizes of relevant individuals
-    relevantIndices <- resolveUniqueEntityIndices relevantEntities indInfoCollection
+    relevantIndices <- resolveUniqueEntityIndices outputOrdered relevantEntities indInfoCollection
 
     -- collect data --
     -- janno