+++ CHANGELOG.md
@@ -1,7 +1,12 @@
-- V 1.5.2.0:
+- V 1.5.3.0:
     - Better error messages for parsec operations on command line input
     - Better error messages for YML file parsing
     - Refactored `--errLength` to exclusively cover exceptions of type `PoseidonGenotypeExceptionForward` emerging during genotype data parsing
+- V 1.5.2.0:
+    - A new option `forge --ordered` was added, which outputs the resulting package with individuals ordered according to the entered entities.
+- V 1.5.1.0:
+    - A new option `list --individuals --fullJanno` adds all standard columns from the Janno to the per-individual output.
+    - A new API option `/individuals?additionalJannoColumns=ALL` triggers the same behaviour for the Web API.
 - V 1.5.0.1: Changed the release pipeline: `trident-macOS` was replaced by `trident-macOS-X64` and `trident-macOS-ARM64`.
 - V 1.5.0.0
     - Removed Josiah Carberry from `newPackageTemplate`, so that he doesn't get added any more to new packages created by `init` and `forge` - the contributor field is missing in the output of these commands now.